refactored the czml conversion helpers inside module
